Relocated help component files and updated snaps
