+++ src/views/Feedback/FeedbackReceived/index.tsx
@@ -2,10 +2,10 @@ import React from 'react';
 import { useTranslation } from 'react-i18next';
 import { Button, GridContainer } from '@trussworks/react-uswds';
 
-import HelpBreadcrumb from 'components/HelpBreadcrumb';
 import UswdsReactLink from 'components/LinkWrapper';
 import MainContent from 'components/MainContent';
 import PageHeading from 'components/PageHeading';
+import HelpBreadcrumb from 'views/HelpAndKnowledge/Articles/_components/HelpBreadcrumb';
 
 const FeedbackReceived = () => {
   const { t } = useTranslation(['feedback']);
