+++ src/views/Feedback/ReportAProblem/index.tsx
@@ -21,12 +21,12 @@ import {
 } from 'gql/gen/graphql';
 
 import BooleanRadio from 'components/BooleanRadioForm';
-import HelpBreadcrumb from 'components/HelpBreadcrumb';
 import MainContent from 'components/MainContent';
 import PageHeading from 'components/PageHeading';
 import Alert from 'components/shared/Alert';
 import FieldGroup from 'components/shared/FieldGroup';
 import { getKeys } from 'types/translation';
+import HelpBreadcrumb from 'views/HelpAndKnowledge/Articles/_components/HelpBreadcrumb';
 
 const ReportAProblem = () => {
   const { t } = useTranslation(['feedback', 'miscellaneous']);
