+++ src/views/HelpAndKnowledge/Articles/AllArticles/index.tsx
@@ -3,9 +3,9 @@ import { useTranslation } from 'react-i18next';
 import { useLocation } from 'react-router-dom';
 import { Grid, GridContainer } from '@trussworks/react-uswds';
 
-import HelpBreadcrumb from 'components/HelpBreadcrumb';
 import MainContent from 'components/MainContent';
 import PageHeading from 'components/PageHeading';
+import HelpBreadcrumb from 'views/HelpAndKnowledge/Articles/_components/HelpBreadcrumb';
 
 import HelpCardGroup from '../_components/HelpCardGroup';
