+++ src/views/HelpAndKnowledge/Articles/_components/RelatedArticles/index.test.tsx
