+++ src/views/HelpAndKnowledge/Articles/_components/RelatedArticles/index.tsx
@@ -4,8 +4,8 @@ import { Link } from 'react-router-dom';
 import { CardGroup, GridContainer } from '@trussworks/react-uswds';
 import classnames from 'classnames';
 
-import ArticleCard from 'components/ArticleCard';
 import helpAndKnowledgeArticles from 'views/HelpAndKnowledge/Articles';
+import ArticleCard from 'views/HelpAndKnowledge/Articles/_components/ArticleCard';
 
 type RelatedArticlesProps = {
   className?: string;
